Reorganize the old 'one-prompt' method into the new, more modulized
workflow.
This PR:
* [x] Merge `generate_targets_for_analysis()` into `WritingStage`, e.g.,
a 'OnePromptPrototyper' class dedicated for this.
* Currently cloud exps fix build errors in the main thread, but will do
it into a cloud build instance like agents
* [x] Merge build step of `check_targets()` into `WritingStage` to
ensure the fuzz target can build `keep the two options:
    * Build_and_run_cloud
    * build_and_run_local

Next PRs:
* [ ] Merge run step of `check_targets()` into `ExecutionStage`, to
collect runtime stats.
* [ ] Merge `triage_crash()` + the following steps into
`AnalysisStage()`->Triager
* [ ] Split `build_and_run_cloud` into 3 functions:
  1. Build and collect results  in 'OnePromptPrototyper'.
  2. Build+run and collect results in `ExecutionStage`.
  3. Triage results in the Triager
* [ ] Split `_fix_generated_fuzz_target()` into 2 functions
  1. Fix build failures in Protyper
  2. Fix runtime failure in Refiner.

+++ run_all_experiments.py
@@ -144,6 +144,7 @@ def run_experiments(benchmark: benchmarklib.Benchmark, args) -> Result:
   """Runs an experiment based on the |benchmark| config."""
   try:
     work_dirs = WorkDirs(os.path.join(args.work_dir, f'output-{benchmark.id}'))
+    args.work_dirs = work_dirs
     model = models.LLM.setup(
         ai_binary=args.ai_binary,
         name=args.model,
@@ -238,8 +239,7 @@ def parse_args() -> argparse.Namespace:
   parser.add_argument(
       '-of',
       '--oss-fuzz-dir',
-      help=
-      'Path to OSS-Fuzz dir to use. If not set will create temporary directory.',
+      help='OSS-Fuzz dir path to use. Create temporary directory by default.',
       default='')
   parser.add_argument(
       '-g',
@@ -547,9 +547,12 @@ def main():
   # Set global variables that are updated throughout experiment runs.
   WORK_DIR = args.work_dir
 
-  coverage_gains_process = Process(
-      target=extend_report_with_coverage_gains_process)
-  coverage_gains_process.start()
+  if args.cloud_experiment_name:
+    coverage_gains_process = Process(
+        target=extend_report_with_coverage_gains_process)
+    coverage_gains_process.start()
+  else:
+    coverage_gains_process = None
 
   experiment_results = []
   if NUM_EXP == 1:
@@ -575,9 +578,10 @@ def main():
       # Wait for all workers to complete.
       p.join()
 
-  # Do a final coverage aggregation.
-  coverage_gains_process.kill()
-  extend_report_with_coverage_gains()
+  if coverage_gains_process:
+    # Do a final coverage aggregation.
+    coverage_gains_process.kill()
+    extend_report_with_coverage_gains()
 
   # Capture time at end
   end = time.time()
